Hide get_syntax, move tests to module
